Added a file accessibility check to `addFile`
I was back and forth on this idea, as throwing an error has its merits, but since we are adding files within a JSON array, if a file is not accessible, giving a `console.error(..)` prompt and not breaking flows better for me. 

Potentially options can be added so throw exceptions if that is more desired by the end user?